Add check_for_tweets daemon
