+++ daemons/check_for_toots.rb
@@ -3,18 +3,7 @@ require 'mastodon_ext'
 require 'mastodon_user_processor'
 require 'toot_transformer'
 require 'httparty'
-
-class InterruptibleSleep
-  def sleep(seconds)
-    @_sleep_check, @_sleep_interrupt = IO.pipe
-    IO.select([@_sleep_check], nil, nil, seconds)
-  end
-
-  def wakeup
-    @_sleep_interrupt.close if @_sleep_interrupt && !@_sleep_interrupt.closed?
-  end
-end
-
+require 'interruptible_sleep'
 
 class CheckForToots
   OLDER_THAN_IN_SECONDS = 30
